backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
ultralytics==11.0.0
opencv-python==4.9.0.80
numpy==1.26.0
streamlit==1.31.0
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.6.0 